Moved calculation of sizes to frontend
