+++ index.js
@@ -3,8 +3,9 @@ const fs = require('fs')
 const express = require('express')
 const app = express()
 const path = require('path')
-const DataCollector = require('./DataCollector')
+const DataCollector = require('./DataCollector');
 const dataCollector = new DataCollector()
+const roomController = require('./RoomController')({dataCollector})
 const WSUpdater = require('js-ws-updater')
 const logger = console
 
@@ -36,7 +37,8 @@ app.get('/persons', jsonService(getNodes))
 
 // app.put('/nodes/:id', (req, res) => res.json(dataCollector.saveNodeChanges(req.params.id, req.body)))
 
-app.post('/rooms', jsonService(req => addRoom(req.body)))
+app.post('/rooms', jsonService(req => roomController.addRoom(req.body)))
+app.put('/rooms', jsonService(req => roomController.addRoom(req.body)))
 
 app.use(express.static(path.join(__dirname, 'public')))
 
@@ -71,42 +73,6 @@ function jsonService (func) {
       next(error)
     }
   }
-}
-
-function prepareTopic(node) {
-  const weight = Math.sqrt(node.weight)
-  node.visibility = weight
-  node.fontSize = Math.sqrt(weight)
-  node.className = 'topic'
-  node.shape = 'rect'
-  node.width = weight * 10 + 50
-  node.height = node.width * 0.7
-}
-
-function preparePerson(node) {
-  node.className = 'person'
-  node.shape = 'circle'
-  node.radius = 50
-}
-
-function prepareRoom(node) {
-  node.className = 'room'
-  node.shape = 'circle'
-  node.radius = 50
-}
-
-function addRoom(data) {
-  data.links = data.links || {}
-  if (data.topics) {
-    data.links.topics = data.topics.map(topic => dataCollector.addNode('topic', topic, prepareTopic))
-    delete data.topics
-  }
-  if (data.users) {
-    data.links.persons = data.users.map(person => dataCollector.addNode('person', person, preparePerson))
-    delete data.users
-  }
-  dataCollector.addNode('room', data, prepareRoom)
-  return {ok: true}
 }
 
 function getNodes() {
